Rework cloning to use forking
Massive reworking of the workflow to clone a repo.

Previously a user needed to invite a staff member to clone the board.

This change uses OAuth to fork the repo to the student’s account,
invite the staff member as a collaborator, accept the invitation to
collaborate, enable issues on the clone, then continue as we did before
with creating the project board and cloning over the columns, cards,
and issues.